pyproject.toml
[tool.poetry]
name = "evaluator"
version = "0.1.0"
description = "Automatic evaluator for basic programming exercises"
authors = ["pablonoya"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
docker = "^5.0.2"
django-rq = "^2.4.1"
mysqlclient = "^2.0.3"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
channels = "^3.0.4"
channels-redis = "^3.3.1"
djangorestframework-simplejwt = "^5.0.0"
factory-boy = "^3.2.1"
django-vite = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
